[ci skip] Checkpoint - initial implementation - restructure crds
